Refactor duktape/scripts. Make new ovms_duktape.h and registry in OvmsDuktape object, shared by OvmsCommand and OvmsScript - to remove the inter-dependency and race conditions
