+++ vehicle/OVMS.V3/components/ovms_script/component.mk
@@ -7,10 +7,14 @@
 # please read the ESP-IDF documents if you need to do this.
 #
 
-COMPONENT_SRCDIRS := src
-COMPONENT_ADD_INCLUDEDIRS := src
-COMPONENT_ADD_LDFLAGS = -Wl,--whole-archive -l$(COMPONENT_NAME) -Wl,--no-whole-archive
 
 ifdef CONFIG_OVMS_SC_JAVASCRIPT_DUKTAPE
+COMPONENT_SRCDIRS := src srcduk
+COMPONENT_ADD_INCLUDEDIRS := src srcduk
 COMPONENT_EMBED_FILES := jsmodembed/pubsub.js jsmodembed/json.js
+else
+COMPONENT_SRCDIRS := src
+COMPONENT_ADD_INCLUDEDIRS := src
 endif
+
+COMPONENT_ADD_LDFLAGS = -Wl,--whole-archive -l$(COMPONENT_NAME) -Wl,--no-whole-archive
